LogApp/requirements.txt
Django==3.2.7
django-3-jet==1.0.7
djangorestframework==3.11.2
Pygments==2.7.4
Markdown==3.2.1
coreapi==2.3.3
psycopg2-binary==2.8.4
dj-database-url==0.5.0
gunicorn==20.0.4
whitenoise==5.0.1
PyYAML==5.4